Create TxStateManager to share
